Added FixtureType enumeration
Added an enumeration, entitled FixtureType, for managing forces
between objects later on.  The only types so far are
footboxSensor, playerBody, masslessBody, and massiveBody (which
does not imply a huge body, but any non-player body which has a
gravitational pull).

+++ mainmenustate.hxx
@@ -12,7 +12,7 @@ class MainMenuState : public State
 {
     private:
         std::vector<sf::Text> displayItems;
-        int current;
+        unsigned short int current;
 
     public:
         MainMenuState(Core* game);
